dusseldorf/zentralbibliothek/requirements.txt
attrs==25.1.0
build==1.2.2.post1
cachetools==5.5.1
colorama==0.4.6
dnspython==2.7.0
idna==3.10
pymongo==4.11.1
python-dateutil==2.9.0.post0
six==1.17.0
